+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Improve subscriber error handler
+
 ### 0.7.0 - 2020-07-11
 
  - Add support for pubsub protocol